+++ packages/utils/src/index.ts
@@ -104,7 +104,7 @@ export {
 } from './lib/reports/generate-md-reports-diff.js';
 export { loadReport } from './lib/reports/load-report.js';
 export { logStdoutSummary } from './lib/reports/log-stdout-summary.js';
-export { scoreReport, scoreAuditsWithTarget } from './lib/reports/scoring.js';
+export { scoreAuditsWithTarget, scoreReport } from './lib/reports/scoring.js';
 export { sortReport } from './lib/reports/sorting.js';
 export type {
   ScoredCategoryConfig,
@@ -142,6 +142,7 @@ export type {
   ExtractArray,
   ExtractArrays,
   ItemOrArray,
+  LooseAutocomplete,
   Prettify,
   WithRequired,
 } from './lib/types.js';